WebServer

+++ kafka-stream/src/main/scala/org/esgi/project/streaming/models/View.scala
@@ -3,7 +3,7 @@ package org.esgi.project.streaming.models
 import play.api.libs.json.{Json, OFormat}
 
 case class View(
-                  _id: Int,
+                  _id: Long,
                   title: String,
                   view_category: String
                 )
